breaking out nersc functions from api.jl
